added user verification for modification/deletion

+++ app/controllers/enrollments_controller.rb
@@ -9,16 +9,15 @@ class EnrollmentsController < ApplicationController
     @enrollment.attended_event_id = params[:event_id]
 
     if @enrollment.save
-      puts "*****************************"
       redirect_to controller: :events, action: :show, id: params[:event_id], notice: "You are now attending this event"
     else
-      puts "#############################"
       redirect_to controller: :events, action: :show, id: params[:event_id], notice: "Something went wrong"
     end
   end
 
   # DELETE /enrollments/1 or /enrollments/1.json
   def destroy
+    redirect_to controller: :events, action: :show, id: params[:event_id], notice: "You can't remove a different user from this event!" unless @enrollment.attendee_id == current_user.id
     if @enrollment.destroy
       redirect_to controller: :events, action: :show, id: params[:event_id], notice: "You are no longer attending this event"
     else
